refactor: Controller 삭제
